cleaned up dataloader_ryan_template, created testing/visualization file, and put plotting functions into km_utils
